+++ tests/modules/test_blazeMeterUploader.py
@@ -1,17 +1,18 @@
+import json
 import logging
 import math
 import os
 import shutil
-from io import BytesIO
 import time
+from io import BytesIO
 
+import bzt.modules.blazemeter
 from bzt.modules.aggregator import DataPoint, KPISet
-from tests import BZTestCase, random_datapoint, __dir__
-from bzt.six import URLError, iteritems, viewvalues
 from bzt.modules.blazemeter import BlazeMeterUploader, BlazeMeterClient, BlazeMeterClientEmul, ResultsFromBZA
 from bzt.modules.blazemeter import MonitoringBuffer
+from bzt.six import URLError, iteritems, viewvalues
+from tests import BZTestCase, random_datapoint, __dir__
 from tests.mocks import EngineEmul
-import bzt.modules.blazemeter
 
 
 class TestBlazeMeterUploader(BZTestCase):
@@ -56,11 +57,23 @@ class TestBlazeMeterUploader(BZTestCase):
         obj.startup()
         obj.engine.stopping_reason = ValueError('wrong value')
         obj.aggregated_second(random_datapoint(10))
-        obj.kpi_buffer[-1][DataPoint.CUMULATIVE][''][KPISet.ERRORS] = [
-            {'msg': 'Forbidden', 'cnt': 7373, 'type': KPISet.ERRTYPE_ASSERT, 'urls': [], KPISet.RESP_CODES: '403'},
-            {'msg': 'Allowed', 'cnt': 7373, 'type': KPISet.ERRTYPE_ERROR, 'urls': [], KPISet.RESP_CODES: '403'}]
+        obj.kpi_buffer[-1][DataPoint.CURRENT][''][KPISet.ERRORS] = [
+            {'msg': 'Forbidden', 'cnt': 10, 'type': KPISet.ERRTYPE_ASSERT, 'urls': [], KPISet.RESP_CODES: '111'},
+            {'msg': 'Allowed', 'cnt': 20, 'type': KPISet.ERRTYPE_ERROR, 'urls': [], KPISet.RESP_CODES: '222'}]
         obj.post_process()
         self.assertEqual(0, len(client.results))
+        data = json.loads(client.requests[6]['data'])
+        self.assertEqual(1, len(data['labels']))
+        total_item = data['labels'][0]
+        self.assertEqual('ALL', total_item['name'])
+        self.assertEqual(total_item['assertions'], [{
+            'failureMessage': 'Forbidden',
+            'failures': 10,
+            'name': 'All Assertions'}])
+        self.assertEqual(total_item['errors'], [{
+            'm': 'Allowed',
+            'count': 20,
+            'rc': '222'}])
 
     def test_check(self):
         client = BlazeMeterClientEmul(logging.getLogger(''))
@@ -122,6 +135,7 @@ class TestBlazeMeterUploader(BZTestCase):
         obj.shutdown()
         log_file = obj.engine.create_artifact('log', '.tmp')
         obj.engine.log.parent.handlers.append(logging.FileHandler(log_file))
+        obj.engine.config.get('modules').get('shellexec').get('env')['TAURUS_INDEX_ALL'] = 1
         obj.post_process()
         self.assertEqual(0, len(client.results))
 
@@ -341,4 +355,3 @@ class TestMonitoringBuffer(BZTestCase):
             mon_buffer.record_data(mon)
         unpacked = sum(item['interval'] for item in viewvalues(mon_buffer.data['local']))
         self.assertEqual(unpacked, ITERATIONS)
-