+++ site/dat/docs/Changelog.md
@@ -6,6 +6,7 @@
  - do not copy Selenium scripts in artifacts dir before running them
  - fix KPI merging error
  - introduce `language` option for Selenium tests
+ - migrate to new online reporting format
  - fix nose plugin crash when setUp/setUpClass raises an exception 
  - add `action` blocks to JMeter's requests
  - fix percentiles values for cloud-based tests
@@ -13,6 +14,7 @@
  - add stubs for `screenshoter` and `capturehar` services
  - add RSpec tests runner for Selenium
  - add Mocha-based tests runner for Selenium
+ - fix artifact upload in the cloud
 
 ## 1.6.8 <sup>1 sep 2016</sup>
  - fix hamcrest installation for Java-based Selenium tests